stream train test works
update pir

bug fixes

bug fixes to vec bank address can be scalar const

bug fixes to vec bank address can be scalar const

+++ src/spatial/lib/HostML.scala
@@ -80,7 +80,7 @@ trait HostML {
   }
 
   def unstaged_sigmoid(x:Float) = {
-    1 / (1 + math.exp(-x)) 
+    1 / (1 + math.exp(-x).toFloat) 
   }
 
 }
